Move collector code to collector.rs
